+++ sekoia_automation/asset_connector/models/connector.py
@@ -1,4 +1,3 @@
-
 from pydantic import BaseModel
 
 from sekoia_automation.asset_connector.models.ocsf.device import DeviceOCSFModel
@@ -8,12 +7,7 @@ from sekoia_automation.asset_connector.models.ocsf.vulnerability import (
     VulnerabilityOCSFModel,
 )
 
-AssetItem = (
-    VulnerabilityOCSFModel|
-    DeviceOCSFModel|
-    UserOCSFModel|
-    SoftwareOCSFModel
-)
+AssetItem = VulnerabilityOCSFModel | DeviceOCSFModel | UserOCSFModel | SoftwareOCSFModel
 
 
 class DefaultAssetConnectorConfiguration(BaseModel):