+++ CHANGELOG.md
@@ -7,11 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## 1.19.3 - 2025-08-04
+## 1.19.3 - 2025-09-01
+
+### Changed
+
+- Add a default logo in the template used to generate new modules
+- Remove `[]` around `Unreleased` entry in the changelog
+- Include CONFIGURE.md into the generated documentation of automation modules
 
 ### Fixed
 
 - Add users fields to enrichment 
+- Fix asset connector post url 
+- Update OCSF fields for software, user and vulnerability
+- Fix some typos
+- Add some extra tests for updating connector fields
 
 ## 1.19.2 - 2025-07-31
 
@@ -36,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix the pydantic import in models.py of the template used to generate new modules
+- Fix the typing of the `send_event` method in the `Trigger` class
 
 ## 1.18.4